fix(scout-for-lol): satisfy matched-betting quality gates
Why

The full PR verification exposed a backend max-lines violation and two whole-repository formatting failures that focused staged-file checks did not cover.

What

Move the anonymous open-market aggregate query into a focused betting module, update its consumers, and normalize the two sweep-file endings flagged by Prettier.

Verification

- bunx turbo run test typecheck lint --filter=@scout-for-lol/backend --filter=@scout-for-lol/data
- bunx prettier --check packages/scout-for-lol/packages/backend/src
- bunx lefthook run pre-commit

No live Discord or production settlement check was run.
