+++ packages/scout-for-lol/packages/backend/src/betting/parlay-sweep.ts
@@ -180,4 +180,3 @@ export async function voidStaleParlayMarkets(
   }
   return count;
 }
-
