+++ packages/scout-for-lol/packages/backend/src/betting/sweep.ts
@@ -437,4 +437,3 @@ export async function closeBettingPoolById(
     requireExpired: false,
   });
 }
-
